Attendance Using QR-Code SEM VI/requirements.txt
qrcode
opencv-python
numpy
streamlit
streamlit_option_menu
datetime
Pillow
